move stuff out of sklearn_compat
